moves camera_geometry to rr_platform to fix circular dependency in build
