--- rr_common/src/camera_geometry/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(rr_camera_geometry CameraGeometry.cpp)
-target_link_libraries(rr_camera_geometry ${catkin_LIBRARIES})
