+++ rr_platform/CMakeLists.txt
@@ -22,7 +22,6 @@ find_package(
     std_msgs
     message_generation
     geometry_msgs
-    rr_common
 )
 
 find_package(OpenCV REQUIRED)
