+++ rr_platform/include/CameraGeometry.h
