+++ rr_platform/include/angle_utils.hpp
