+++ rr_platform/src/camera/CameraGeometry.cpp
@@ -1,10 +1,10 @@
-#include "rr_common/CameraGeometry.h"
+#include "CameraGeometry.h"
 
 #include <cmath>
 #include <tf/transform_listener.h>
 #include <tf/transform_datatypes.h>
 
-#include "rr_common/angle_utils.hpp"
+#include "angle_utils.hpp"
 
 
 namespace rr {
